backend/requirements.txt
aiomysql==0.3.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.0.1
certifi==2026.4.22
cffi==2.0.0
click==8.3.3
colorama==0.4.6
cryptography==48.0.0
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.136.1
greenlet==3.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.15
joblib==1.5.3
limits==5.8.0
numpy==2.4.4
packaging==26.2
pandas==3.0.3
passlib==1.7.4
playwright==1.59.0
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.4
pydantic-settings==2.14.1
pydantic_core==2.46.4
pyee==13.0.1
PyJWT==2.12.1
PyMySQL==1.1.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.28
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.49
starlette==1.0.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
uvicorn==0.46.0
wrapt==2.1.2
